pyproject.toml
[project]
name = "fastapi-jinja22"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic==1.16.5",
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "bcrypt==4.0.1",
    "certifi==2025.8.3",
    "cffi==2.0.0",
    "click==8.2.1",
    "colorama==0.4.6",
    "cryptography==46.0.1",
    "dnspython==2.8.0",
    "dotenv==0.9.9",
    "email-validator==2.3.0",
    "fastapi==0.116.2",
    "fastapi-cli==0.0.11",
    "fastapi-cloud-cli==0.1.5",
    "greenlet==3.2.4",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "iniconfig==2.1.0",
    "jinja2==3.1.6",
    "mako==1.3.10",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "packaging==25.0",
    "passlib==1.7.4",
    "pluggy==1.6.0",
    "psycopg==3.2.10",
    "psycopg-binary==3.2.10",
    "pycparser==2.23",
    "pydantic==2.11.9",
    "pydantic-core==2.33.2",
    "pygments==2.19.2",
    "pyjwt==2.10.1",
    "pytest==8.4.2",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "rich==14.1.0",
    "rich-toolkit==0.15.1",
    "rignore==0.6.4",
    "sentry-sdk==2.38.0",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.43",
    "starlette==0.48.0",
    "typer==0.17.4",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "watchfiles==1.1.0",
    "websockets==15.0.1",
]
